requirements.txt
django
djangorestframework
coverage
django-cors-headers
django-filter
